pyproject.toml
[build-system]
requires = ["setuptools>=61.0.0", "setuptools-scm>=8.0"]
build-backend = "setuptools.build_meta"

[project]
classifiers = [
    "Intended Audience :: Developers",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
name = "nomad-catalysis"
description = "A NOMAD plugin for heterogeneous catalysis data."
dynamic = ["version"]
readme = "README.md"
requires-python = ">=3.9"
authors = [
    { name = "Julia Schumann"},
    { name = "Hampus Näsström"},
    { name = "Michael Götte"},
    { email = "fairmat@physik.hu-berlin.de"},
]
maintainers = [
    { name = "Julia Schumann", email = "jschumann@physik.hu-berlin.de" },
    { name = "FAIRmat", email = "fairmat@physik.hu-berlin.de"},
]
license = { file = "LICENSE" }
dependencies = ["nomad-lab>=1.3.16", "openpyxl"]

[project.urls]
Repository = "https://github.com/FAIRmat-NFDI/nomad-catalysis-plugin"

[project.optional-dependencies]
dev = ["ruff", "pytest", "structlog"]

[tool.ruff]
# Exclude a variety of commonly ignored directories.
exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".git-rewrite",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".pytype",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

# Same as Black.
line-length = 88
indent-width = 4


[tool.ruff.lint]
select = [
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # isort
    "I",
    # pylint
    "PL",
]

ignore = [
    "F403", # 'from module import *' used; unable to detect undefined names
]

fixable = ["ALL"]

# Allow unused variables when underscore-prefixed.
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"

# this is entirely optional, you can remove this if you wish to
[tool.ruff.format]
# use single quotes for strings.
quote-style = "single"

# indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"

[tool.setuptools]
package-dir = { "" = "src" }

[tool.setuptools.packages.find]
where = ["src"]

[project.entry-points.'nomad.plugin']

catalysis = "nomad_catalysis.schema_packages:catalysis"
catalysis_app = "nomad_catalysis.apps:catalysis"
catalysis_parser = "nomad_catalysis.parsers:catalysis"
catalysis_collection = "nomad_catalysis.parsers:catalysis_collection"
example_catalysis = "nomad_catalysis.example_uploads:catalysis"

[tool.cruft]
# Avoid updating workflow files, this leads to permissions issues
skip = [".github/*"]

[tool.setuptools_scm]
